+++ src/compiler/logic/scope-variable-declaration.ts
@@ -5,18 +5,43 @@
  * @since 0.1.0
  */
 import { KipperStorageType, KipperType } from "./types";
-import { CompoundStatement, VariableDeclaration } from "../tokens";
+import {
+	CompoundStatement,
+	Declaration,
+	FunctionDefinition,
+	ParameterDeclaration,
+	VariableDeclaration,
+} from "../tokens";
 import type { KipperProgramContext } from "../program-ctx";
 
 /**
- * Represents the definition of a scope entry that may be a child of the global scope, a function scope or compound
- * statement scope
+ * Abstract class as a parent for {@link ScopeVariableDeclaration} and {@link Scope}.
+ * @since 0.1.2
  */
-export class ScopeVariableDeclaration {
+export abstract class ScopeDeclaration {
+	public abstract get token(): Declaration;
+	public abstract get identifier(): string;
+
+	/**
+	 * Fetches the {@link KipperProgramContext program context instance} for this token.
+	 */
+	public get programCtx(): KipperProgramContext {
+		return this.token.programCtx;
+	}
+}
+
+/**
+ * Represents the definition of a scope entry that may be a child of the global scope, a function scope or compound
+ * statement scope.
+ * @since 0.1.0
+ */
+export class ScopeVariableDeclaration extends ScopeDeclaration {
 	public constructor(
 		// eslint-disable-next-line no-unused-vars
 		private _token: VariableDeclaration,
-	) {}
+	) {
+		super();
+	}
 
 	/**
 	 * Returns the {@link VariableDeclaration token} this scope declaration bases on.
@@ -35,14 +60,14 @@ export class ScopeVariableDeclaration {
 	/**
 	 * The variable type or return type of this scope entry
 	 */
-	public get type(): KipperType | undefined {
+	public get type(): KipperType {
 		return this._token.valueType;
 	}
 
 	/**
 	 * The storage type of this scope entry
 	 */
-	public get storageType(): KipperStorageType | undefined {
+	public get storageType(): KipperStorageType {
 		return this._token.storageType;
 	}
 
@@ -51,5 +76,48 @@ export class ScopeVariableDeclaration {
 	 */
 	public get scope(): KipperProgramContext | CompoundStatement {
 		return this._token.scope;
+	}
+}
+
+/**
+ * Represents the definition of a function inside a {@link KipperProgramContext program}.
+ * @since 0.1.2
+ */
+export class ScopeFunctionDeclaration extends ScopeDeclaration {
+	public constructor(
+		// eslint-disable-next-line no-unused-vars
+		private _token: FunctionDefinition,
+		// eslint-disable-next-line no-unused-vars
+		private _programCtx: KipperProgramContext,
+	) {
+		super();
+	}
+
+	/**
+	 * Returns the {@link FunctionDefinition token} this scope function declaration bases on.
+	 */
+	public get token(): FunctionDefinition {
+		return this._token;
+	}
+
+	/**
+	 * The identifier of this entry.
+	 */
+	public get identifier(): string {
+		return this._token.identifier;
+	}
+
+	/**
+	 * The function return type.
+	 */
+	public get returnType(): KipperType | undefined {
+		return this._token.returnType;
+	}
+
+	/**
+	 * The storage type of this scope entry.
+	 */
+	public get args(): Array<ParameterDeclaration> | undefined {
+		return this._token.args;
 	}
 }